слияние веток

+++ main.cpp
@@ -5,8 +5,7 @@ using namespace std;
 int main()
 {
     cout << "Enter A and B: ";
-
-int a, b, max, min;
+double a, b, max, min;
 cin >> a >> b;
 cout << "A + B = " << a + b << '\n';
 cout << "A - B = " << a - b << '\n';
